Refactor user module to align with the recommended project tree for clean file structure to begin docker infrastructure,  removed old user files, added modules directory, updated app module
