+++ src/modules/users/controllers/users.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get, Param, Post, Patch, Delete, Body, Query, ParseIntPipe, ValidationPipe } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
+import { UsersService } from '../services/users.service';
+import { CreateUserDto } from '../dto/create-user.dto';
+import { UpdateUserDto } from '../dto/update-user.dto';
 
 @Controller('users')
 export class UsersController {
@@ -38,4 +38,4 @@ export class UsersController {
     delete(@Param('id', ParseIntPipe) id: number) {
         return this.usersService.delete(id);
     }
-}
+}
