--- src/users/users.service.ts
@@ -1,120 +1,0 @@
-import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
-import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
-
-@Injectable()
-export class UsersService {
-  private users = [
-    {
-      id: 1,
-      name: "Leanne Graham",
-      email: "Sincere@april.biz",
-      role: "INTERN",
-    },
-    {
-      id: 2,
-      name: "Ervin Howell",
-      email: "Shanna@melissa.tv",
-      role: "ADMIN",
-    },
-    {
-      id: 3,
-      name: "Clementine Bauch",
-      email: "Nathan@yesenia.net",
-      role: "USER",
-    },
-    {
-      id: 4,
-      name: "Patricia Lebsack",
-      email: "Julianne.OConner@kory.org",
-      role: "USER",
-    },
-    {
-      id: 5,
-      name: "Chelsey Dietrich",
-      email: "Lucio_Hettinger@annie.ca",
-      role: "INTERN",
-    },
-    {
-      id: 6,
-      name: "Mrs. Dennis Schulist",
-      email: "Karley_Dach@jasper.info",
-      role: "USER",
-    },
-    {
-      id: 7,
-      name: "Kurtis Weissnat",
-      email: "Telly.Hoeger@billy.biz",
-      role: "MODERATOR",
-    },
-    {
-      id: 8,
-      name: "Nicholas Runolfsdottir V",
-      email: "Sherwood@rosamond.me",
-      role: "ADMIN",
-    },
-    {
-      id: 9,
-      name: "Glenna Reichert",
-      email: "Chaim_McDermott@dana.io",
-      role: "USER",
-    },
-    {
-      id: 10,
-      name: "Clementina DuBuque",
-      email: "Rey.Padberg@karina.biz",
-      role: "INTERN",
-    },
-  ];
-
-  findAll(role?: 'INTERN' | 'ADMIN' | 'ENGINEER') {
-    if (role) {
-      const rolesArray = this.users.filter((user) => user.role.toLowerCase() === role.toLowerCase());
-      if (rolesArray.length === 0)
-        throw new NotFoundException('user role not found');
-      return rolesArray;
-    }
-    return this.users;
-  }
-
-  findOne(id: number) {
-    const user = this.users.find((user) => user.id === id);
-    if (!user) {
-      throw new NotFoundException('User not found');
-    }
-    return this.users.find(user => user.id === id);
-  }
-
-  create(createUserDto: CreateUserDto) {
-    const userByHighestId = [...this.users].sort((a, b) => b.id - a.id)[0];
-    const newUser = {
-      id: userByHighestId.id + 1,
-      ...createUserDto,
-    };
-    this.users.push(newUser);
-    return newUser;
-  }
-
-  update(id: number, updateUserDto: UpdateUserDto) {
-    this.users = this.users.map((user) => {
-      if (user.id === id) {
-        return {
-          ...user,
-          ...updateUserDto,
-        }
-      }
-      return user;
-    })
-    return this.findOne(id); // return the updated user with the updated id
-  }
-
-  delete(id: number) {
-    const removedUser = this.findOne(id);
-    if (!removedUser) {
-      throw new NotFoundException('User not found');
-    }
-    this.users = this.users.filter((user) => user.id !== id); // exclude the removed user
-    return removedUser;
-  }
-
-}
